+++ api/test/tests/e2e/track_upload/test_upload_track_with_fingerprinting_failure.py
@@ -38,7 +38,7 @@ class TestCase(UploadedTrackTestCase):
             if fingerprint_missing_cause:
                 code = fingerprint_missing_cause.code.code
                 assert code in [
-                    FingerprintMissingCauseCode.Codes.AUDIO_META_AMALYSIS_DISABLED,
+                    FingerprintMissingCauseCode.Codes.AFP_DISABLED,
                     FingerprintMissingCauseCode.Codes.SERVICE_NOT_FOUND,
                     FingerprintMissingCauseCode.Codes.WRONG_FILE_EXTENSION,
                     FingerprintMissingCauseCode.Codes.WRONG_FILE_TYPE,