+++ api/test/tests/integration/uploaded_track/input/method/post/attributes/file/fingerprint/test_missing_cause_not_enabled.py
@@ -23,4 +23,4 @@ class TestCase(UploadedTrackTestCase):
         assert self.saved_object.track_file
         assert self.saved_object.track_file.fingerprint_missing_cause
         assert self.saved_object.track_file.fingerprint_missing_cause.code.code == \
-            FingerprintMissingCauseCode.Codes.AUDIO_META_AMALYSIS_DISABLED
+            FingerprintMissingCauseCode.Codes.AFP_DISABLED