+++ api/test/tests/integration/uploaded_track/input/method/post/attributes/file/musicbrainz_recording/test_musicbrainz_recording_missing_cause_meta_analyse_disabled.py
@@ -1,4 +1,4 @@
-import pytest
+from django.test import override_settings
 from rest_framework import status
 
 from api.model.musicbrainz_resource.children.recording.missing_cause.code.MbRecordingMissingCauseCode import (
@@ -8,12 +8,12 @@ from api.test.utils.uploaded_track.UploadedTrackTestFilename import UploadedTrac
 from api.test.tests.integration.uploaded_track.UploadedTrackTestCase import UploadedTrackTestCase
 
 
-@pytest.mark.parametrize('enable_audio_metadata_analysis', [False], indirect=True)
 class TestCase(UploadedTrackTestCase):
-    def test_audio_meta_analysis_disabled_then_corresponding_missing_cause(self, enable_audio_metadata_analysis):
-        response = self._post_uploaded_track(UploadedTrackTestFilename.RECORDING_SHOWMUSTGOON_MP3)
+    def test_audio_meta_analysis_disabled_then_corresponding_missing_cause(self):
+        with override_settings(AFP_ENABLED=False):
+            response = self._post_uploaded_track(UploadedTrackTestFilename.RECORDING_SHOWMUSTGOON_MP3)
 
         assert response.status_code == status.HTTP_201_CREATED
         assert self.saved_object.track_file.musicbrainz_recording_missing_cause
         assert self.saved_object.track_file.musicbrainz_recording_missing_cause.code.code == \
-            MbRecordingMissingCauseCode.Codes.AUDIO_META_AMALYSIS_DISABLED
+            MbRecordingMissingCauseCode.Codes.AFP_DISABLED